User_login and register fixed

+++ app/models/participant_models.py
@@ -1,5 +1,5 @@
 # app/models/participant_models.py
-from sqlalchemy import Column, String, DateTime,Integer,Boolean
+from sqlalchemy import Column, String, DateTime, Boolean, BigInteger
 from sqlalchemy.orm import declarative_base
 
 Base = declarative_base()
@@ -12,5 +12,5 @@ class Participant(Base):
     email = Column(String(100), unique=True, index=True)
     password = Column(String(200),nullable=False)
     created_at = Column(DateTime)
-    studentID = Column(Integer, unique=True, nullable=False)
+    studentID = Column(BigInteger, unique=True, nullable=False)
     leader = Column(Boolean, default= False)
